Change int to float

+++ src/arduino/smartbridge/src/component/Lcd.h
@@ -7,7 +7,8 @@ class Lcd {
     public:
         virtual void on() = 0;
         virtual void off() = 0;
-        virtual void writePreAlarm(String msg, int waterLevel) = 0;
-        virtual void writeAlarm(String msg, int waterLevel, int angle) = 0;
+        virtual void writePreAlarm(String msg, float waterLevel) = 0;
+        virtual void writeAlarm(String msg, float waterLevel, int angle) = 0;
 };
+
 #endif
